+++ queries/models/monitoramento/veiculo_dia.sql
@@ -50,15 +50,15 @@ with
                     }}
                 )
                 or (
-                    data between "2025-07-16" and "2025-07-31" -- Excessão para lacres adicionados após o prazo em 2025-07-Q2
+                    data between "2025-07-16" and "2025-07-31"  -- Exceção para lacres adicionados após o prazo em 2025-07-Q2
                     and data_processamento between "2025-07-16" and "2025-08-13"
                 )
                 or (
-                    data between "2025-08-01" and "2025-08-18" -- Excessão para falha do arquivo de licenciamento no ftp
+                    data between "2025-08-01" and "2025-08-18"  -- Exceção para falha do arquivo de licenciamento no ftp
                     and data_processamento between "2025-08-01" and "2025-08-26"
                 )
                 or (
-                    data between "2025-09-01" and "2025-09-18" -- Exceção para lacres adicionados após o prazo em 2025-09-Q1
+                    data between "2025-09-01" and "2025-09-18"  -- Exceção para lacres adicionados após o prazo em 2025-09-Q1
                     and data_processamento between "2025-09-01" and "2025-09-18"
                 )
             )